Move engine api base class to alosi lib
